getUserRoles separate file
